move serve http code to a different file
